+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+- provide meaningful Exception when custom Enum is implemented incorrectly.
+
 ## [0.1.0] – 2016-09-09
 ### Added
 - abstract class Enum with unit tests